pyproject.toml
[project]
name = "qr-supercharge"
version = "0.1.0"
description = "QR code generator with embedded pixel-art text"
requires-python = ">=3.10"
dependencies = [
    "qrcode[pil]>=7.4",
    "Pillow>=10.0",
    "pyzbar>=0.1.9",
    "click>=8.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black",
    "ruff",
]

[project.scripts]
qr-embed = "qr_supercharge.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qr_supercharge"]
